Update of Strategy class
I fixed errors in the code. The only problem remaining is with the get_data function: it doesn't recognize the functions from the Stock's class. I am still trying to fix it but I commit so the rest of the team can work with the updated version of the others functions with the error fixed.