Add drafts of StrategyDashboard and StrategyGraph classes
I have realised that calling a simulation function from a Strategy object is not exactly what we want. I rather wrote a StrategyDashBoard Class which has as data member a list of StrategyGraph objects which are QCustomPlot objects. The user might be able to create a limited number of StrategyDashBoard tabs at a time. Each tab contains a StrategyGraph object which can be initialized so as to display a strategy for many stocks or many stocks for a strategy or many stocks for many strategies. Recall that a Strategy class is initialized with a stock, a data frequency (ByMinute or ByDay) and a price type ("o","h","l" or "c").